Implement UpdateMirrors as Symfony command
