+++ src/AppBundle/Command/Update/UpdateCommand.php
@@ -2,12 +2,11 @@
 
 namespace AppBundle\Command\Update;
 
-use archportal\cronjobs\UpdateMirrors;
 use archportal\cronjobs\UpdateNews;
 use archportal\cronjobs\UpdatePackages;
-use archportal\cronjobs\UpdatePkgstats;
 use archportal\cronjobs\UpdateReleases;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Command\LockableTrait;
 use Symfony\Component\Console\Exception\InvalidArgumentException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,6 +14,8 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class UpdateCommand extends ContainerAwareCommand
 {
+    use LockableTrait;
+
     protected function configure()
     {
         $this
@@ -24,13 +25,11 @@ class UpdateCommand extends ContainerAwareCommand
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $this->lock('cron.lock', true);
         $this->getContainer()->get('AppBundle\Service\LegacyEnvironment')->initialize();
 
         $job = $input->getArgument('job');
         switch ($job) {
-            case 'mirrors':
-                UpdateMirrors::run();
-                break;
             case 'news':
                 UpdateNews::run();
                 break;
